Add disclosure example
